TELEGRAM/requirements.txt
aiogram==2.25.1
asyncpg==0.27.0
python-dotenv==1.0.0
uvloop==0.17.0
